[misc] clean up user initialization in app (#1654)
* [misc] clean up

* handle validate error

* fix ci

* simplify validation error

---------
