+++ ymir/backend/src/ymir_app/app/db/init_db.py
@@ -5,10 +5,7 @@ from sqlalchemy.orm import Session
 
 from app import crud, schemas
 from app.config import settings
-from app.constants.role import Roles
 from app.db import base  # noqa: F401
-from app.utils.security import frontend_hash
-from app.utils.ymir_controller import ControllerClient
 
 
 # make sure all SQL Alchemy models are imported (app.db.base) before initializing DB
@@ -21,26 +18,6 @@ def init_db(db: Session) -> None:
     # But if you don't want to use migrations, create
     # the tables un-commenting the next line
     # Base.metadata.create_all(bind=engine)
-    roles = crud.role.get_multi(db)
-    if not roles:
-        for role in [Roles.NORMAL, Roles.ADMIN, Roles.SUPER_ADMIN]:
-            role_in = schemas.RoleCreate(name=role.name, description=role.description)
-            crud.role.create(db, obj_in=role_in)
-
-    user = crud.user.get_by_email(db, email=settings.FIRST_ADMIN)
-    if not user:
-        password = frontend_hash(settings.FIRST_ADMIN_PASSWORD)
-        user_in = schemas.UserCreate(
-            email=settings.FIRST_ADMIN,
-            password=password,
-        )
-        user = crud.user.create(db, obj_in=user_in)  # noqa: F841
-        user = crud.user.activate(db, user=user)
-        user = crud.user.update_role(db, user=user, role=schemas.UserRole.SUPER_ADMIN)
-
-        if settings.INIT_LABEL_FOR_FIRST_USER:
-            controller = ControllerClient(settings.GRPC_CHANNEL)
-            controller.create_user(user_id=user.id)
 
     docker_images = crud.docker_image.get_multi(db)
     if not docker_images and settings.DOCKER_IMAGES:
